competitive-intel/requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data validation
pydantic>=2.5.0

# Text processing
html2text>=2020.1.16

# LLM generation
anthropic>=0.40.0

# Embeddings (Anthropic-compatible via Voyager or local)
# For Anthropic embedding usage, use the anthropic SDK directly
# sentence-transformers>=2.0.0  # Optional: local embeddings

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
tenacity>=8.2.0

# Deduplication
datasketch>=1.6.0

# Date parsing
python-dateutil>=2.8.0

# Fast JSON serialization
orjson>=3.9.0
